Feat: COX-3 added role
