--- apps/api/src/app/modules/role/seeds/role.seed.ts
@@ -1,17 +1,0 @@
-export const ROLE_SEED: any = [
-  {
-    name: 'SuperAdmin',
-    permissions: 'all',
-    isModifiable: false,
-  },
-  {
-    name: 'Admin',
-    permissions: 'none',
-    isModifiable: true,
-  },
-  {
-    name: 'User',
-    permissions: 'none',
-    isModifiable: true,
-  },
-];
