+++ apps/api/src/app/modules/user/seeds/role.seed.ts
@@ -1,0 +1,26 @@
+export const ROLE_SEED: any = [
+  {
+    name: 'SuperAdmin',
+    permissions: {
+      description: 'permission1',
+      active: true,
+    },
+    isModifiable: false,
+  },
+  {
+    name: 'Admin',
+    permissions: {
+      description: 'permission1',
+      active: true,
+    },
+    isModifiable: true,
+  },
+  {
+    name: 'Collaborator',
+    permissions: {
+      description: 'permission1',
+      active: false,
+    },
+    isModifiable: true,
+  },
+];
