+++ apps/api/src/app/modules/user/services/role.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { CrudService } from '@shared/classes/crud-service.class';
-import { Role } from '@modules/role/entities/role.entity';
-import { RoleCreateDtoV } from '@modules/role/dtov/role-create.dtov';
-import { RoleUpdateDtoV } from '@modules/role/dtov/role-update.dtov';
+import { Role } from '@modules/user/entities/role.entity';
+import { RoleCreateDtoV } from '@modules/user/dtov/role-create.dtov';
+import { RoleUpdateDtoV } from '@modules/user/dtov/role-update.dtov';
 
 @Injectable()
 export class RoleService extends CrudService<Role, RoleCreateDtoV, RoleUpdateDtoV> {
