+++ libs/api-interfaces/src/index.ts
@@ -2,4 +2,5 @@ export * from './interfaces/base-response.interface';
 export * from './interfaces/find-boosted.interface';
 
 export * from './lib/api-interfaces';
-export * from "./dto/role/role.dto";
+export * from "./dto/user/user.dto";
+export * from "./dto/user/role.dto";
